Copies yontma to %PROGRAMFILES%\yontma\yontma.exe before installing the service so that we're not relying on the file to always stay in the originally launched location (which may just be a downloads foldeR).
Still needs some code cleanup.

Partial fix for issue #11. though not a complete fix as this code has a minor bug. If the installer is launced in x86 mode, yontma will be installed to Program Files (x86) on an x64 machine, but if the uninstaller is run in x64 mode, it will look in the wrong place for the yontma binaries. Better solution is to get the path of the binary from service manager.